requirements.txt
asgiref==3.3.1
Django==3.1.5
Pillow==8.1.0
pypng==0.0.20
PyQRCode==1.2.1
pytz==2018.7
sqlparse==0.4.1
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.11.2
idna==2.7
Markdown==3.0.1
oauthlib==2.1.0
python3-openid==3.1.0
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.2

